refactor(provider): move passport projector to application layer
