+++ backend/src/main/java/com/alligator/market/backend/provider/application/passport/projection/ProviderPassportProjector.java
@@ -1,4 +1,4 @@
-package com.alligator.market.domain.provider.readmodel.passport.projection.projector;
+package com.alligator.market.backend.provider.application.passport.projection;
 
 import com.alligator.market.domain.provider.passport.ProviderPassport;
 import com.alligator.market.backend.provider.application.passport.projection.port.out.ProviderPassportProjectionWritePort;
